store/wizard: move details selectors to submodule
