+++ src/store/slices/wizard/details/index.ts
@@ -1,3 +1,4 @@
+export * from './selectors';
 export { initialState as detailsState } from './state';
 export * from './types';
 export * from './utilities';
